go/consensus/tendermint: Move log adapter to common package
